requirements.txt
Django>=4.2.1,<4.3
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.6,<2.10
drf-spectacular>=0.26.2,<0.27